refactor: rename s3Client to configClient
