+++ packages/api/src/config.ts
@@ -1,0 +1,1 @@
+export * from "./config/index";
