+++ packages/api/src/config/browser.ts
@@ -1,9 +1,9 @@
 import ky, { type Options } from "ky";
 
-import { S3Client } from "./isomorphic";
+import { ConfigClient } from "./isomorphic";
 
 export const createS3BrowserClient = (options: Options) =>
-  S3Client.init({
+  ConfigClient.init({
     target: "browser",
     instance: ky.extend(options),
   });
