--- packages/api/src/s3.ts
@@ -1,1 +1,0 @@
-export * from "./s3/index";
